A bunch of changes to fix compiler warnings in Xcode 4.5.

+++ source/infrastructure/MUReadBuffer.h
@@ -15,7 +15,7 @@
 - (NSUInteger) length;
 
 - (NSData *) dataByConsumingBuffer;
-- (NSData *) dataByConsumingBytesToIndex: (unsigned) index;
+- (NSData *) dataByConsumingBytesToIndex: (NSUInteger) index;
 - (NSData *) dataValue;
 
 - (NSString *) ASCIIStringByConsumingBuffer;
